20231110-product-004 tests 데칼코마니로 나눴습니다. DB는 FK 적용했습니다.
